file-server-manager/requirements.txt
flask==0.12.1
python-socketio==1.7.4
eventlet==0.21.0